requirements.txt
pyswisseph>=2.10.0
rich>=13.0.0
geopy>=2.4.0
